Adding Cluster Capabilities
A basic cluster is implemented using [hashicorp/memberlist](https://github.com/hashicorp/memberlist)

Cluster uses redis to store inflight and retain messages and subscriptions across nodes.
