+++ server/persistence/bolt/bolt.go
@@ -2,6 +2,7 @@ package bolt
 
 import (
 	"errors"
+	"strconv"
 	"time"
 
 	sgob "github.com/asdine/storm/codec/gob"
@@ -18,6 +19,7 @@ const (
 
 var (
 	errNotFound = "not found"
+	errNotOpened = errors.New("boltdb not opened")
 )
 
 // Store is a backend for writing and reading to bolt persistent storage.
@@ -59,6 +61,18 @@ func (s *Store) Open() error {
 // Close closes the boltdb instance.
 func (s *Store) Close() {
 	s.db.Close()
+}
+
+func (s *Store ) GenSubscriptionId(cid, filter string) string {
+	return persistence.KSubscription + "_" + cid + ":" + filter
+}
+
+func (s *Store ) GenInflightId(cid string, pid uint16) string {
+	return persistence.KInflight + "_" + cid + "_" + strconv.Itoa(int(pid))
+}
+
+func (s *Store ) GenRetainedId(topic string) string {
+	return persistence.KRetained + "_" + topic
 }
 
 // WriteServerInfo writes the server info to the boltdb instance.
@@ -127,13 +141,12 @@ func (s *Store) WriteClient(v persistence.Client) error {
 }
 
 // DeleteSubscription deletes a subscription from the boltdb instance.
-func (s *Store) DeleteSubscription(id string) error {
+func (s *Store) DeleteSubscription(cid, filter string) error {
 	if s.db == nil {
 		return errors.New("boltdb not opened")
 	}
-
 	err := s.db.DeleteStruct(&persistence.Subscription{
-		ID: id,
+		ID: s.GenSubscriptionId(cid, filter),
 	})
 	if err != nil {
 		return err
@@ -159,13 +172,13 @@ func (s *Store) DeleteClient(id string) error {
 }
 
 // DeleteInflight deletes an inflight message from the boltdb instance.
-func (s *Store) DeleteInflight(id string) error {
+func (s *Store) DeleteInflight(cid string, pid uint16) error {
 	if s.db == nil {
 		return errors.New("boltdb not opened")
 	}
 
 	err := s.db.DeleteStruct(&persistence.Message{
-		ID: id,
+		ID: s.GenInflightId(cid, pid),
 	})
 	if err != nil {
 		return err
@@ -175,13 +188,13 @@ func (s *Store) DeleteInflight(id string) error {
 }
 
 // DeleteRetained deletes a retained message from the boltdb instance.
-func (s *Store) DeleteRetained(id string) error {
+func (s *Store) DeleteRetained(topic string) error {
 	if s.db == nil {
 		return errors.New("boltdb not opened")
 	}
 
 	err := s.db.DeleteStruct(&persistence.Message{
-		ID: id,
+		ID: s.GenRetainedId(topic),
 	})
 	if err != nil {
 		return err
@@ -193,7 +206,7 @@ func (s *Store) DeleteRetained(id string) error {
 // ReadSubscriptions loads all the subscriptions from the boltdb instance.
 func (s *Store) ReadSubscriptions() (v []persistence.Subscription, err error) {
 	if s.db == nil {
-		return v, errors.New("boltdb not opened")
+		return v, errNotOpened
 	}
 
 	err = s.db.Find("T", persistence.KSubscription, &v)
@@ -207,7 +220,7 @@ func (s *Store) ReadSubscriptions() (v []persistence.Subscription, err error) {
 // ReadClients loads all the clients from the boltdb instance.
 func (s *Store) ReadClients() (v []persistence.Client, err error) {
 	if s.db == nil {
-		return v, errors.New("boltdb not opened")
+		return v, errNotOpened
 	}
 
 	err = s.db.Find("T", persistence.KClient, &v)
@@ -221,7 +234,7 @@ func (s *Store) ReadClients() (v []persistence.Client, err error) {
 // ReadInflight loads all the inflight messages from the boltdb instance.
 func (s *Store) ReadInflight() (v []persistence.Message, err error) {
 	if s.db == nil {
-		return v, errors.New("boltdb not opened")
+		return v, errNotOpened
 	}
 
 	err = s.db.Find("T", persistence.KInflight, &v)
@@ -235,7 +248,7 @@ func (s *Store) ReadInflight() (v []persistence.Message, err error) {
 // ReadRetained loads all the retained messages from the boltdb instance.
 func (s *Store) ReadRetained() (v []persistence.Message, err error) {
 	if s.db == nil {
-		return v, errors.New("boltdb not opened")
+		return v, errNotOpened
 	}
 
 	err = s.db.Find("T", persistence.KRetained, &v)
@@ -249,10 +262,66 @@ func (s *Store) ReadRetained() (v []persistence.Message, err error) {
 //ReadServerInfo loads the server info from the boltdb instance.
 func (s *Store) ReadServerInfo() (v persistence.ServerInfo, err error) {
 	if s.db == nil {
-		return v, errors.New("boltdb not opened")
+		return v, errNotOpened
 	}
 
 	err = s.db.One("ID", persistence.KServerInfo, &v)
+	if err != nil && err.Error() != errNotFound {
+		return
+	}
+
+	return v, nil
+}
+
+//ReadSubscriptionsByCid loads all the subscriptions from the boltdb instance by the client id.
+func (s *Store) ReadSubscriptionsByCid(cid string) (v []persistence.Subscription, err error) {
+	if s.db == nil {
+		return v, errNotOpened
+	}
+
+	err = s.db.Find("Client", cid, &v)
+	if err != nil && err.Error() != errNotFound {
+		return
+	}
+
+	return v, nil
+}
+
+//ReadInflightByCid loads all the inflight messages from the boltdb instance by the client id.
+func (s *Store) ReadInflightByCid(cid string) (v []persistence.Message, err error) {
+	if s.db == nil {
+		return v, errNotOpened
+	}
+
+	err = s.db.Find("Client", cid, &v)
+	if err != nil && err.Error() != errNotFound {
+		return
+	}
+
+	return v, nil
+}
+
+//ReadRetainedByTopic loads the retained message from the boltdb instance by the topic.
+func (s *Store) ReadRetainedByTopic(topic string) (v persistence.Message, err error) {
+	if s.db == nil {
+		return v, errNotOpened
+	}
+
+	err = s.db.One("ID", s.GenRetainedId(topic), &v)
+	if err != nil && err.Error() != errNotFound {
+		return
+	}
+
+	return v, nil
+}
+
+//ReadClientByCid loads the client info from the boltdb instance.
+func (s *Store) ReadClientByCid(cid string) (v persistence.Client, err error) {
+	if s.db == nil {
+		return v, errNotOpened
+	}
+
+	err = s.db.One("ClientID", cid, &v)
 	if err != nil && err.Error() != errNotFound {
 		return
 	}
